refactor(atomix): transform data carriers to records
Transform several Raft data classes to Java records. This is mostly a
maintenance change to cut down on boilerplate, and not for performance
reasons.
