+++ atomix/cluster/src/main/java/io/atomix/raft/storage/log/entry/ApplicationEntry.java
@@ -15,10 +15,7 @@
  */
 package io.atomix.raft.storage.log.entry;
 
-import static com.google.common.base.MoreObjects.toStringHelper;
-
 import java.nio.ByteBuffer;
-import java.util.Objects;
 import org.agrona.DirectBuffer;
 import org.agrona.concurrent.UnsafeBuffer;
 
@@ -27,62 +24,11 @@ import org.agrona.concurrent.UnsafeBuffer;
  * and highestPosition metadata allow for fast binary search over a collection of entries to quickly
  * find a particular record.
  */
-public class ApplicationEntry implements RaftEntry {
-
-  private final long lowestPosition;
-  private final long highestPosition;
-  private final DirectBuffer data = new UnsafeBuffer();
+public record ApplicationEntry(long lowestPosition, long highestPosition, DirectBuffer data)
+    implements RaftEntry {
 
   public ApplicationEntry(
       final long lowestPosition, final long highestPosition, final ByteBuffer data) {
-    this.lowestPosition = lowestPosition;
-    this.highestPosition = highestPosition;
-    this.data.wrap(data);
-  }
-
-  public ApplicationEntry(
-      final long lowestPosition, final long highestPosition, final DirectBuffer data) {
-    this.lowestPosition = lowestPosition;
-    this.highestPosition = highestPosition;
-    this.data.wrap(data);
-  }
-
-  public long lowestPosition() {
-    return lowestPosition;
-  }
-
-  public long highestPosition() {
-    return highestPosition;
-  }
-
-  public DirectBuffer data() {
-    return data;
-  }
-
-  @Override
-  public String toString() {
-    return toStringHelper(this)
-        .add("lowestPosition", lowestPosition())
-        .add("highestPosition", highestPosition())
-        .toString();
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final ApplicationEntry that = (ApplicationEntry) o;
-    return lowestPosition == that.lowestPosition
-        && highestPosition == that.highestPosition
-        && data.equals(that.data);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(lowestPosition, highestPosition, data);
+    this(lowestPosition, highestPosition, new UnsafeBuffer(data));
   }
 }
